requirements.txt
--extra-index-url https://download.pytorch.org/whl/nightly/cpu
torch 
biopython
tqdm
ipykernel
matplotlib
matplotlib-inline
einops
datasets
fair-esm
